backend/requirements.txt
fastapi
uvicorn[standard]
neo4j
pysus
psutil
hypothesis
pytest
pytest-asyncio
python-dotenv
httpx
google-genai
groq
openpyxl
xlrd
pandas
scipy
